Move Gltf type definition into load_gltf
