+++ src/factory/load_gltf.rs
@@ -16,16 +16,68 @@ use mint;
 use std::{fs, io};
 
 use camera::Camera;
-use gltf::Gltf;
 use gltf_utils::AccessorIter;
 use object::Object;
 use skeleton::Skeleton;
 use std::path::{Path, PathBuf};
 use vec_map::VecMap;
 
+use animation::Clip;
 use {Group, Material, Mesh, Texture};
 use geometry::{Geometry, Shape};
+use object;
 use super::Factory;
+
+/// Loaded glTF 2.0 returned by [`Factory::load_gltf`].
+///
+/// [`Factory::load_gltf`]: struct.Factory.html#method.load_gltf
+#[derive(Debug, Clone)]
+pub struct Gltf {
+    /// Imported camera views.
+    pub cameras: Vec<Camera>,
+
+    /// Imported animation clips.
+    pub clips: Vec<Clip>,
+
+    /// The node heirarchy of the default scene.
+    ///
+    /// If the `glTF` contained no default scene then this
+    /// container will be empty.
+    pub heirarchy: VecMap<object::Group>,
+
+    /// Imported mesh instances.
+    ///
+    /// ### Notes
+    ///
+    /// * Must be kept alive in order to be displayed.
+    pub instances: Vec<Mesh>,
+
+    /// Imported mesh materials.
+    pub materials: Vec<Material>,
+
+    /// Imported mesh templates.
+    pub meshes: VecMap<Vec<Mesh>>,
+
+    /// The root node of the default scene.
+    ///
+    /// If the `glTF` contained no default scene then this group
+    /// will have no children.
+    pub root: object::Group,
+
+    /// Imported skeletons.
+    pub skeletons: Vec<Skeleton>,
+
+    /// Imported textures.
+    pub textures: Vec<Texture<[f32; 4]>>,
+}
+
+impl AsRef<object::Base> for Gltf {
+    fn as_ref(&self) -> &object::Base {
+        self.root.as_ref()
+    }
+}
+
+impl object::Object for Gltf {}
 
 fn build_scene_hierarchy(
     factory: &mut Factory,
@@ -159,7 +211,7 @@ fn load_textures(
 }
 
 fn load_materials(
-    gltf: &Gltf,
+    gltf: &gltf::Gltf,
     textures: &[Texture<[f32; 4]>],
 ) -> Vec<Material> {
     let mut materials = Vec::new();
@@ -329,7 +381,7 @@ fn load_meshes(
 
 fn load_skeletons(
     factory: &mut Factory,
-    gltf: &Gltf,
+    gltf: &gltf::Gltf,
     heirarchy: &VecMap<Group>,
     buffers: &gltf_importer::Buffers,
 ) -> Vec<Skeleton> {
@@ -365,7 +417,7 @@ fn load_skeletons(
 }
 
 fn load_clips(
-    gltf: &Gltf,
+    gltf: &gltf::Gltf,
     heirarchy: &VecMap<Group>,
     buffers: &gltf_importer::Buffers,
 ) -> Vec<animation::Clip> {
@@ -484,7 +536,7 @@ impl super::Factory {
     pub fn load_gltf(
         &mut self,
         path_str: &str,
-    ) -> super::Gltf {
+    ) -> Gltf {
         info!("Loading {}", path_str);
         let path = Path::new(path_str);
         let base = path.parent().unwrap_or(&Path::new(""));
@@ -511,7 +563,7 @@ impl super::Factory {
             instances = bind_objects(self, &gltf, &heirarchy, &meshes, &skeletons);
         }
 
-        super::Gltf {
+        Gltf {
             cameras,
             clips,
             heirarchy,
